--- a/CPT_DAPT_QLoRA_minimal.ipynb
+++ b/CPT_DAPT_QLoRA_minimal.ipynb
@@ -7,7 +7,7 @@
    "source": [
     "\n",
     "# 🚀 CPT/DAPT + Tiny Anchor with QLoRA (Minimal Notebook)\n",
-    "This notebook scaffolds **Domain-Adaptive Pretraining (CPT/DAPT)** on your DPIP docs using **QLoRA**, \n",
+    "This notebook scaffolds **Domain-Adaptive Pretraining (CPT/DAPT)** on your docs using **QLoRA**, \n",
     "mixed with a **small instruction anchor (5–10%)** to preserve instruction-following.\n",
     "\n",
     "**Steps**: chunk docs → build CPT dataset → (what is a collator) → interleave CPT+anchor → load model → train → guardrails → eval.\n"
@@ -36,7 +36,7 @@
     "from peft import LoraConfig, get_peft_model, prepare_model_for_kbit_training\n",
     "\n",
     "CFG = {\n",
-    "    \"run_name\": \"dapt_gemma3_dpip_v0\",\n",
+    "    \"run_name\": \"dapt_gemma3_v0\",\n",
     "    \"model_name\": \"google/gemma-3-27b-it\",  # or base if available\n",
     "    \"block_size\": 2048,\n",
     "    \"pack_factor\": 4,\n",
@@ -54,8 +54,8 @@
     "    \"use_bf16\": True,\n",
     "    \"gradient_checkpointing\": True,\n",
     "    \"output_dir\": \"./outputs_dapt\",\n",
-    "    \"dpip_raw_dir\": \"./data/raw\",\n",
-    "    \"dpip_cpt_jsonl\": \"./data/processed/dpip_cpt.jsonl\",\n",
+    "    \"raw_dir\": \"./data/raw\",\n",
+    "    \"cpt_jsonl\": \"./data/processed/cpt.jsonl\",\n",
     "    \"anchor_jsonl\": \"./data/processed/anchor_instr.jsonl\",\n",
     "    \"lora_r\": 16,\n",
     "    \"lora_alpha\": 32,\n",
@@ -131,13 +131,13 @@
     "    print(f\"Wrote {count} CPT chunks -> {out_jsonl}\")\n",
     "\n",
     "# Demo file if nothing exists\n",
-    "if not Path(CFG[\"dpip_cpt_jsonl\"]).exists():\n",
-    "    os.makedirs(CFG[\"dpip_raw_dir\"], exist_ok=True)\n",
+    "if not Path(CFG[\"cpt_jsonl\"]).exists():\n",
+    "    os.makedirs(CFG[\"raw_dir\"], exist_ok=True)\n",
     "    with open(\"./data/raw/demo.txt\", \"w\") as f:\n",
-    "        f.write((\"DPIP Screening Service...\\n\" * 50))\n",
-    "    make_cpt_jsonl_from_dir(CFG[\"dpip_raw_dir\"], CFG[\"dpip_cpt_jsonl\"])\n",
+    "        f.write((\"Screening Service...\\n\" * 50))\n",
+    "    make_cpt_jsonl_from_dir(CFG[\"raw_dir\"], CFG[\"cpt_jsonl\"])\n",
     "else:\n",
-    "    print(\"Found CPT jsonl:\", CFG[\"dpip_cpt_jsonl\"])\n"
+    "    print(\"Found CPT jsonl:\", CFG[\"cpt_jsonl\"])\n"
    ]
   },
   {
@@ -151,7 +151,7 @@
     "# --- Tiny anchor (Alpaca-like) ---\n",
     "def write_demo_anchor(out_jsonl):\n",
     "    demo = [\n",
-    "        {\"instruction\": \"Explain the DPIP Screening Service in one paragraph.\", \"input\":\"\", \"output\":\"It performs low-latency fraud checks by querying a local cache...\"},\n",
+    "        {\"instruction\": \"Explain the Screening Service in one paragraph.\", \"input\":\"\", \"output\":\"It performs low-latency fraud checks by querying a local cache...\"},\n",
     "        {\"instruction\": \"List two benefits of local cache screening.\", \"input\":\"\", \"output\":\"Lower latency and resilience to transient central outages.\"}\n",
     "    ]\n",
     "    with open(out_jsonl, \"w\", encoding=\"utf-8\") as f:\n",
@@ -223,7 +223,7 @@
     "    tokenized = tokenized.map(lambda e: group_texts(e, CFG[\"block_size\"]), batched=True)\n",
     "    return tokenized\n",
     "\n",
-    "cpt_ds = load_cpt_dataset(CFG[\"dpip_cpt_jsonl\"], block_size=CFG[\"block_size\"])\n",
+    "cpt_ds = load_cpt_dataset(CFG[\"cpt_jsonl\"], block_size=CFG[\"block_size\"])\n",
     "anchor_ds = load_anchor_dataset(CFG[\"anchor_jsonl\"])\n",
     "train_ds = interleave_datasets([cpt_ds, anchor_ds], probabilities=[CFG[\"cpt_weight\"], CFG[\"anchor_weight\"]], seed=CFG[\"seed\"])\n",
     "\n",
@@ -365,12 +365,12 @@
     "    ppl = math.exp(float(np.mean(nlls)))\n",
     "    return ppl\n",
     "\n",
-    "# Example (replace with held-out DPIP paragraphs):\n",
-    "# dev_texts = [\"Held-out DPIP paragraph ...\", \"Another paragraph ...\"]\n",
+    "# Example (replace with held-out paragraphs):\n",
+    "# dev_texts = [\"Held-out paragraph ...\", \"Another paragraph ...\"]\n",
     "# print(\"Perplexity:\", perplexity_on_texts(model, tokenizer, dev_texts))\n",
     "\n",
     "def run_rag_eval_stub(checkpoint_dir: str):\n",
-    "    # Integrate your DPIP RAG eval harness here.\n",
+    "    # Integrate your  RAG eval harness here.\n",
     "    return {\"faithfulness\": None, \"completeness\": None, \"specificity\": None, \"conciseness\": None, \"schema_pass_rate\": None}\n"
    ]
   }