… ACCELERATE_SETUP_GUIDE.md
… CHANGELOG.md
… GETTING_STARTED.md
… MULTI_GPU_GUIDE.md
… README.md
… Trails/train_distributed_v1.py
… Trails/train_distributed_v2.py
… VERSION
… configs/Samples/run_distributed_v1.yaml
… configs/deepseed_z2.json
… configs/deepspeed_z2.json
… configs/run_bnb.yaml
… configs/run_dapt.yaml
… configs/run_distributed.yaml
… configs/run_unsloth.yaml
… configs/test_rendered.yaml
… data-generation/README.md
… pyproject.toml
… requirements.txt
… scripts/train_distributed.py
… tests/merge_lora_merge_infer.py
… tests/test_evaluation_optimization.py